+++ internal/user/handler.go
@@ -35,7 +35,7 @@ func (h *handler) ListUsers(c *echo.Context) error {
 		return response.Error(c, err)
 	}
 
-	return response.WithPagination(c, http.StatusOK, users, params.Page, params.Limit, int64(total))
+	return response.WithPagination(c, http.StatusOK, users, params.Page, params.Limit, total)
 }
 
 func (h *handler) CreateUser(c *echo.Context) error {
